pyproject.toml
[project]
name = "miniforge"
version = "0.1.0"
description = "High-performance MiniMax M2.7 inference library optimized for GMKtech M7"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Miniforge User"},
]
keywords = ["llm", "inference", "minimax", "gguf", "quantization", "amd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "psutil>=5.9",
    "pillow>=10.0",
    "huggingface-hub>=0.20",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "typing-extensions>=4.8",
    "numpy>=1.24",
    "llama-cpp-python>=0.3.20",
]

[project.optional-dependencies]
llama-cpp = [
    "llama-cpp-python>=0.3.5",
]
server = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "starlette>=0.36.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.2.0",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
all = [
    "miniforge[server,dev]",
]

[project.urls]
Homepage = "https://github.com/user/miniforge"
Repository = "https://github.com/user/miniforge"
Issues = "https://github.com/user/miniforge/issues"

[project.scripts]
miniforge = "miniforge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.2.0",
    "mypy>=1.8",
]

[tool.uv.pip]
extra-index-url = [
    "https://download.pytorch.org/whl/cpu",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warnings
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "N818",  # exception names
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
show_error_codes = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/miniforge"]

[tool.hatch.build]
dev-mode-dirs = ["src"]
